+++ brownie/project/main.py
@@ -103,7 +103,8 @@ class _ProjectBase:
             os.chdir(cwd)
 
         for alias, data in build_json.items():
-            if self._build_path is not None:
+            if self._build_path is not None and not data["sourcePath"].startswith("interface"):
+                # interfaces should generate artifact in /build/interfaces/ not /build/contracts/
                 if alias == data["contractName"]:
                     # if the alias == contract name, this is a part of the core project
                     path = self._build_path.joinpath(f"contracts/{alias}.json")
@@ -651,7 +652,7 @@ def compile_source(
     if vyper_version is None:
         # if no vyper compiler version is given, try to compile using solidity
         compiler_config["solc"] = {
-            "version": solc_version or str(compiler.solidity.get_version()),
+            "version": solc_version or str(compiler.solidity.get_version().truncate()),
             "optimize": optimize,
             "runs": runs,
         }